module reorganisation #395
