+++ applications/add_file_to_db.py
@@ -37,7 +37,7 @@
 import logging
 from pathlib import Path
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import db_handler
