+++ applications/db_development_tools/add_new_parameter_to_db.py
@@ -13,7 +13,7 @@
 import logging
 
 import simtools.config as cfg
-import simtools.util.commandline_parser as argparser
+import simtools.configuration.commandline_parser as argparser
 import simtools.util.general as gen
 from simtools import db_handler
 from simtools.util import names
