+++ applications/get_file_from_db.py
@@ -31,7 +31,7 @@
 
 import logging
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import db_handler
