+++ applications/make_regular_arrays.py
@@ -29,7 +29,7 @@ from pathlib import Path
 
 import astropy.units as u
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import db_handler, io_handler
 from simtools.layout.layout_array import LayoutArray
