+++ applications/plot_simtel_histograms.py
@@ -32,7 +32,7 @@ from pathlib import Path
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools.simtel.simtel_histograms import SimtelHistograms
