+++ applications/print_array_elements.py
@@ -40,7 +40,7 @@
 import logging
 from pathlib import Path
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools.layout import layout_array
