+++ applications/produce_array_config.py
@@ -85,7 +85,7 @@
 import logging
 from pathlib import Path
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools.model.array_model import ArrayModel
