+++ applications/production.py
@@ -57,7 +57,7 @@ from copy import copy
 
 from astropy.io.misc import yaml
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools.simulator import Simulator
