+++ applications/sim_showers_for_trigger_rates.py
@@ -57,7 +57,7 @@ from pathlib import Path
 
 import astropy.units as u
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import io_handler
 from simtools.simulator import Simulator
