+++ applications/validate_camera_efficiency.py
@@ -46,7 +46,7 @@
 import logging
 from pathlib import Path
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import io_handler
 from simtools.camera_efficiency import CameraEfficiency
