+++ applications/validate_camera_fov.py
@@ -44,7 +44,7 @@
 import logging
 from pathlib import Path
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import io_handler
 from simtools.model.telescope_model import TelescopeModel
