+++ applications/validate_optics.py
@@ -65,7 +65,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.backends.backend_pdf import PdfPages
 
-import simtools.configuration as configurator
+import simtools.configuration.configurator as configurator
 import simtools.util.general as gen
 from simtools import io_handler
 from simtools.model.telescope_model import TelescopeModel
