+++ simtools/configuration/commandline_parser.py
