+++ simtools/data_model/validate_data.py
