+++ simtools/data_model/validate_schema.py
@@ -2,7 +2,7 @@ import datetime
 import logging
 import re
 
-import simtools.util.data_model as data_model
+import simtools.data_model.data_model as data_model
 import simtools.util.general as gen
