+++ simtools/model/model_utils.py
