+++ simtools/ray_tracing.py
@@ -13,11 +13,11 @@ from astropy.table import QTable
 
 import simtools.util.general as gen
 from simtools import io_handler, visualize
+from simtools.model.model_utils import compute_telescope_transmission
 from simtools.model.telescope_model import TelescopeModel
 from simtools.psf_analysis import PSFImage
 from simtools.simtel.simtel_runner_ray_tracing import SimtelRunnerRayTracing
 from simtools.util import names
-from simtools.util.model import compute_telescope_transmission
 
 __all__ = ["RayTracing"]
