+++ simtools/simulator.py
@@ -10,7 +10,7 @@ import numpy as np
 import simtools.util.general as gen
 from simtools import io_handler
 from simtools.corsika.corsika_runner import CorsikaRunner
-from simtools.job_submission.job_manager import JobManager
+from simtools.job_execution.job_manager import JobManager
 from simtools.model.array_model import ArrayModel
 from simtools.simtel.simtel_histograms import SimtelHistograms
 from simtools.simtel.simtel_runner_array import SimtelRunnerArray
