+++ tests/unit_tests/job_execution/test_job_manager.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-import simtools.job_submission.job_manager as jm
+import simtools.job_execution.job_manager as jm
 
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
